nit: added baseStyle to same file
